Rename fsm.h -> power.h
