+++ alpha/power.h
