+++ alpha/ui.cpp
@@ -1,6 +1,6 @@
 #include "ui.h"
 #include "hal.h"
-#include "fsm.h"
+#include "power.h"
 
 
 // Increase as UI thread uses more things
